+++ src/types.ts
@@ -1,3 +1,0 @@
-export type CssRuleDefinition = [string, string[]];
-
-export type CssRuleDefinitionSet = CssRuleDefinition[];